Sample - BasicMvvmApp renamed to `SampleMvvmApp`
