+++ samples/SampleMvvmApp/SampleMvvmApp.csproj
