+++ samples/SampleMvvmApp/Views/DashboardView.xaml
@@ -5,7 +5,7 @@
              xmlns:prism="http://prismlibrary.com/"
              prism:ViewModelLocator.AutoWireViewModel="True"
              mc:Ignorable="d" d:DesignWidth="500" d:DesignHeight="350"
-             x:Class="BasicMvvmApp.Views.DashboardView">
+             x:Class="SampleMvvmApp.Views.DashboardView">
 
   <UserControl.Styles>
     <!--<Style Selector="Button.cmdBtn">-->
