+++ samples/SampleMvvmApp/Views/DashboardView.xaml.cs
@@ -1,7 +1,7 @@
 ﻿using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace BasicMvvmApp.Views
+namespace SampleMvvmApp.Views
 {
     /// <summary>DashboardView.</summary>
     public partial class DashboardView : UserControl
