+++ samples/SampleMvvmApp/Views/MainWindow.xaml
@@ -2,11 +2,11 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:core="clr-namespace:BasicMvvmApp;assembly=BasicMvvmApp"
-        xmlns:views="using:BasicMvvmApp.Views"
+        xmlns:core="clr-namespace:SampleMvvmApp;assembly=SampleMvvmApp"
+        xmlns:views="using:SampleMvvmApp.Views"
         xmlns:prism="http://prismlibrary.com/"
         prism:ViewModelLocator.AutoWireViewModel="True"
-        x:Class="BasicMvvmApp.Views.MainWindow"
+        x:Class="SampleMvvmApp.Views.MainWindow"
         Title="{Binding Title}"
         Height="500" Width="700"
         TransparencyLevelHint="AcrylicBlur"
