+++ samples/SampleMvvmApp/Views/MainWindow.xaml.cs
@@ -2,7 +2,7 @@
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace BasicMvvmApp.Views
+namespace SampleMvvmApp.Views
 {
     public partial class MainWindow : Window
     {
