+++ samples/SampleMvvmApp/Views/SettingsView.xaml.cs
@@ -1,7 +1,7 @@
 ﻿using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace BasicMvvmApp.Views
+namespace SampleMvvmApp.Views
 {
     /// <summary>Sample Settings View.</summary>
     public partial class SettingsView : UserControl
