+++ samples/SampleMvvmApp/Views/SubSettingsView.axaml
@@ -5,7 +5,7 @@
              xmlns:prism="http://prismlibrary.com/"
              prism:ViewModelLocator.AutoWireViewModel="True"
              mc:Ignorable="d" d:DesignWidth="500" d:DesignHeight="350"
-             x:Class="BasicMvvmApp.Views.SubSettingsView">
+             x:Class="SampleMvvmApp.Views.SubSettingsView">
   <DockPanel>
     <StackPanel DockPanel.Dock="Top" Spacing="5">
       <Grid Background="LightGray" ColumnDefinitions="Auto, *">
